% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISTMNP_function.R
\name{GISTMNP}
\alias{GISTMNP}
\title{GISTMNP Calculation Function}
\usage{
GISTMNP(num_dim, num_obj, manifold)
}
\arguments{
\item{num_dim}{Number of dimensions defining each stimulus object}

\item{num_obj}{Number of objects belonging to categorical stimulus}

\item{manifold}{Structural Manifold values specified in 0-1 interval}
}
\value{
The GISTMNP prediction given what was specified for the three parameters
}
\description{
The purpose of this function is to calculate GISTMNP predictions given user-specified input
}
\examples{
family_341 <- GISTMNP(3,4,c(1,1,0))
}

